requirements.txt
fastapi
uvicorn
sqlalchemy
GeoAlchemy2
asyncpg
alembic
passlib[bcrypt]
pydantic
python-jose[cryptography]
pyotp
aiosmtplib
python-dotenv
email-validator
motor
httpx
logging
